Create bad_good_but_both_are_ugly.java
this is an example over an actual project to a subject of mine.
this is very bad implementation of the solution and also feels weak.
on the other hand this shows as disregard for Overriding of methods from interface, which in casual java IDE would not want to compile :L 

The problem is obvious, the interface creates methods for setting Author and Text as a string. Programmer is lazy and weird though and wanted to make an amazing overhaul of a code with tinkering with lightweight List of Characters instead of complicated and highlevel String class. Programmer also made some unreasonable decisions which lead to a unreasonable usages of complicated methods in background.
The solution of a programmer was simply abstracting the whole String and List problematics in order to unify these two complex and extremely different approaches. 

Liskov was hereby violated with not overriding methods.
Liskov was hereby solved with another class on top of simple List of Characters or String.